requirements.txt
chainlit
openai
python-dotenv
requests
langchain
langchain-community
langchain-openai
sqlalchemy
psycopg2-binary